Separate time utility functions to a new cli module
